fix(govtask): progresso fisico e financeiro real na lista de processos
Os cards mostravam sempre 0% porque os dois percentuais vinham de campos
que ninguem preenche, ignorando os dados de execucao ja registrados:

- Fisico saia so de obras.percentual_fisico, campo manual (PATCH de obra).
  Passa a cair no acumulado da ultima medicao aprovada/paga quando a obra
  nao tem percentual lancado.
- Financeiro usava convenio.valor_executado, tambem manual. Passa a somar
  os movimentos do tipo PAGAMENTO, a mesma definicao que o dashboard ja
  usava ("Valor executado = soma dos pagamentos"), eliminando a divergencia
  entre as duas telas.

O percentual manual da obra continua tendo precedencia, entao lancamento
explicito segue sobrepondo o valor derivado.

Serializacao: Decimal virava string no JSON ("25.00") enquanto o front tipa
number e renderizava o valor cru. Schema passa a float arredondado a 1 casa
e o card formata escondendo a casa decimal zerada.

Verificado em producao: "Construcao escola" (medicao aprovada de 25% e
pagamento de R$ 250k em R$ 1M) passou de 0%/0.0% para 25%/25%; os demais
seguem em 0% por nao terem medicao nem pagamento.

+++ modulo-govtask/api/app/schemas/convenio.py
@@ -180,8 +180,8 @@ class ConvenioListItem(BaseModel):
     orgao_concedente: str | None = None
     etapa_atual: str | None = None
     proximo_prazo: datetime | None = None
-    percentual_fisico: Decimal | None = None
-    percentual_financeiro: Decimal | None = None
+    percentual_fisico: float | None = None
+    percentual_financeiro: float | None = None
     tarefas_abertas: int = 0
     pendencias: int = 0
     responsavel: dict | None = None
